requirements.txt
PyMuPDF==1.23.7
sentence-transformers
psycopg2-binary
python-dotenv
flask
numpy
tqdm
pandas
faiss-cpu
